Remove intermediate file usage
This makes sense for debugging, but results in a fragile
system when using this in real life. Since we load all the
data into memory in a single chunk anyway, this should result
in no change in performance.

We also implement a single file for storing state including
tokens used when parsing later datasets.